Move CoffeeNetSecurityAutoConfiguration to .servlet
Mirroring `SecurityAutoConfiguration` from Spring Boot, this class will
be used only for servlet based environments. We will later add a
reactive variant.
