+++ coffeenet-autoconfigure/src/test/java/rocks/coffeenet/autoconfigure/security/servlet/DelegatingSecurityConfigurerTests.java
@@ -17,8 +17,6 @@ import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
 
-import rocks.coffeenet.autoconfigure.security.CoffeeNetSecurityAutoConfiguration;
-
 import static org.assertj.core.api.Assertions.assertThat;
